split the classes that are related to temporary file management off from standard buffer manager
